refactor(lib): isolate redaction lexical folding (split S01 L1/3)
